feat(eeprom): Move cartridge into its own module
